+++ src/renderer/src/components/editor/diff-viewer-props.ts
@@ -1,0 +1,28 @@
+import type { LargeDiffRenderLimit } from './large-diff-render-limit'
+
+export type DiffViewerProps = {
+  modelKey: string
+  originalModelKey?: string
+  modifiedModelKey?: string
+  originalContent: string
+  modifiedContent: string
+  language: string
+  filePath: string
+  relativePath: string
+  sideBySide: boolean
+  editable?: boolean
+  // Why: optional because DiffViewer is also used by GitHubItemDialog for PR
+  // review, where there is no local worktree to attach comments to.
+  worktreeId?: string
+  onAddLineComment?: (args: {
+    lineNumber: number
+    startLine?: number
+    body: string
+  }) => Promise<boolean>
+  commentableLineNumbers?: readonly number[]
+  addLineCommentLabel?: string
+  addLineCommentPlaceholder?: string
+  onContentChange?: (content: string) => void
+  onSave?: (content: string) => void
+  largeDiffRenderLimit?: LargeDiffRenderLimit
+}
